renamed views dir because no longer conflict with views.py
