+++ cnxarchive/views/resource.py
