+++ cnxarchive/views/search_view.py
